requirements.txt
requests
flask
bcrypt
pyjwt
flask_cors
python-dotenv
pyodbc